Refactor classification Models
