+++ src/otx/algo/classification/timm_model.py
@@ -1,30 +1,27 @@
-# Copyright (C) 2024 Intel Corporation
+# Copyright (C) 2023 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
-
-"""MobileNetV3 model implementation."""
-
-
+#
+"""EfficientNetV2 model implementation."""
 from __future__ import annotations
 
 from copy import deepcopy
-from typing import TYPE_CHECKING, Any, Literal
+from typing import TYPE_CHECKING, Literal
 
 import torch
-from torch import Tensor, nn
+from torch import nn
 
-from otx.algo.classification.backbones import OTXMobileNetV3
+from otx.algo.classification.backbones.timm import TimmBackbone, TimmModelType
 from otx.algo.classification.classifier import ImageClassifier, SemiSLClassifier
 from otx.algo.classification.heads import (
-    HierarchicalNonLinearClsHead,
+    HierarchicalLinearClsHead,
     LinearClsHead,
-    MultiLabelNonLinearClsHead,
+    MultiLabelLinearClsHead,
     OTXSemiSLLinearClsHead,
 )
 from otx.algo.classification.losses.asymmetric_angular_loss_with_ignore import AsymmetricAngularLossWithIgnore
 from otx.algo.classification.necks.gap import GlobalAveragePooling
 from otx.algo.classification.utils import get_classification_layers
 from otx.algo.utils.support_otx_v1 import OTXv1Helper
-from otx.core.data.entity.base import OTXBatchLossEntity
 from otx.core.data.entity.classification import (
     HlabelClsBatchDataEntity,
     HlabelClsBatchPredEntity,
@@ -33,12 +30,13 @@ from otx.core.data.entity.classification import (
     MultilabelClsBatchDataEntity,
     MultilabelClsBatchPredEntity,
 )
-from otx.core.exporter.base import OTXModelExporter
-from otx.core.exporter.native import OTXNativeModelExporter
-from otx.core.metrics import MetricInput
 from otx.core.metrics.accuracy import HLabelClsMetricCallble, MultiClassClsMetricCallable, MultiLabelClsMetricCallable
 from otx.core.model.base import DefaultOptimizerCallable, DefaultSchedulerCallable
-from otx.core.model.classification import OTXHlabelClsModel, OTXMulticlassClsModel, OTXMultilabelClsModel
+from otx.core.model.classification import (
+    OTXHlabelClsModel,
+    OTXMulticlassClsModel,
+    OTXMultilabelClsModel,
+)
 from otx.core.schedulers import LRSchedulerListCallable
 from otx.core.types.label import HLabelInfo, LabelInfoTypes
 from otx.core.types.task import OTXTrainType
@@ -49,32 +47,22 @@ if TYPE_CHECKING:
     from otx.core.metrics import MetricCallable
 
 
-class MobileNetV3ForMulticlassCls(OTXMulticlassClsModel):
-    """MobileNetV3ForMulticlassCls is a class that represents a MobileNetV3 model for multiclass classification.
-
-    Args:
-        mode (Literal["large", "small"]): The mode of the MobileNetV3 model, either "large" or "small".
-        num_classes (int): The number of classes for classification.
-        loss_callable (Callable[[], nn.Module], optional): The loss function callable. Defaults to nn.CrossEntropyLoss.
-        optimizer (OptimizerCallable, optional): The optimizer callable. Defaults to DefaultOptimizerCallable.
-        scheduler (LRSchedulerCallable | LRSchedulerListCallable, optional): The learning rate scheduler callable.
-            Defaults to DefaultSchedulerCallable.
-        metric (MetricCallable, optional): The metric callable. Defaults to MultiClassClsMetricCallable.
-        torch_compile (bool, optional): Whether to compile the model using TorchScript. Defaults to False.
-        freeze_backbone (bool, optional): Whether to freeze the backbone layers during training. Defaults to False.
-    """
+class TimmModelForMulticlassCls(OTXMulticlassClsModel):
+    """TimmModel for multi-class classification task."""
 
     def __init__(
         self,
         label_info: LabelInfoTypes,
-        mode: Literal["large", "small"] = "large",
+        backbone: TimmModelType = "efficientnetv2_s_21k",
+        pretrained: bool = True,
         optimizer: OptimizerCallable = DefaultOptimizerCallable,
         scheduler: LRSchedulerCallable | LRSchedulerListCallable = DefaultSchedulerCallable,
         metric: MetricCallable = MultiClassClsMetricCallable,
         torch_compile: bool = False,
         train_type: Literal[OTXTrainType.SUPERVISED, OTXTrainType.SEMI_SUPERVISED] = OTXTrainType.SUPERVISED,
     ) -> None:
-        self.mode = mode
+        self.backbone = backbone
+        self.pretrained = pretrained
 
         super().__init__(
             label_info=label_info,
@@ -100,17 +88,16 @@ class MobileNetV3ForMulticlassCls(OTXMulticlassClsModel):
         return model
 
     def _build_model(self, num_classes: int) -> nn.Module:
-        backbone = OTXMobileNetV3(mode=self.mode)
+        backbone = TimmBackbone(backbone=self.backbone, pretrained=self.pretrained)
         neck = GlobalAveragePooling(dim=2)
         loss = nn.CrossEntropyLoss(reduction="none")
-        in_channels = 960 if self.mode == "large" else 576
         if self.train_type == OTXTrainType.SEMI_SUPERVISED:
             return SemiSLClassifier(
                 backbone=backbone,
                 neck=neck,
                 head=OTXSemiSLLinearClsHead(
                     num_classes=num_classes,
-                    in_channels=in_channels,
+                    in_channels=backbone.num_features,
                     loss=loss,
                 ),
             )
@@ -120,7 +107,7 @@ class MobileNetV3ForMulticlassCls(OTXMulticlassClsModel):
             neck=neck,
             head=LinearClsHead(
                 num_classes=num_classes,
-                in_channels=in_channels,
+                in_channels=backbone.num_features,
                 topk=(1, 5) if num_classes >= 5 else (1,),
                 loss=loss,
             ),
@@ -128,7 +115,7 @@ class MobileNetV3ForMulticlassCls(OTXMulticlassClsModel):
 
     def load_from_otx_v1_ckpt(self, state_dict: dict, add_prefix: str = "model.") -> dict:
         """Load the previous OTX ckpt according to OTX2.0."""
-        return OTXv1Helper.load_cls_mobilenet_v3_ckpt(state_dict, "multiclass", add_prefix)
+        return OTXv1Helper.load_cls_effnet_v2_ckpt(state_dict, "multiclass", add_prefix)
 
     def forward_explain(self, inputs: MulticlassClsBatchDataEntity) -> MulticlassClsBatchPredEntity:
         """Model forward explain function."""
@@ -144,7 +131,7 @@ class MobileNetV3ForMulticlassCls(OTXMulticlassClsModel):
             feature_vector=outputs["feature_vector"],
         )
 
-    def forward_for_tracing(self, image: Tensor) -> Tensor | dict[str, Tensor]:
+    def forward_for_tracing(self, image: torch.Tensor) -> torch.Tensor | dict[str, torch.Tensor]:
         """Model forward function used for the model tracing during model exportation."""
         if self.explain_mode:
             return self.model(images=image, mode="explain")
@@ -152,19 +139,22 @@ class MobileNetV3ForMulticlassCls(OTXMulticlassClsModel):
         return self.model(images=image, mode="tensor")
 
 
-class MobileNetV3ForMultilabelCls(OTXMultilabelClsModel):
-    """MobileNetV3 Model for multi-class classification task."""
+class TimmModelForMultilabelCls(OTXMultilabelClsModel):
+    """TimmModel for multi-label classification task."""
 
     def __init__(
         self,
         label_info: LabelInfoTypes,
-        mode: Literal["large", "small"] = "large",
+        backbone: TimmModelType = "efficientnetv2_s_21k",
+        pretrained: bool = True,
         optimizer: OptimizerCallable = DefaultOptimizerCallable,
         scheduler: LRSchedulerCallable | LRSchedulerListCallable = DefaultSchedulerCallable,
         metric: MetricCallable = MultiLabelClsMetricCallable,
         torch_compile: bool = False,
     ) -> None:
-        self.mode = mode
+        self.backbone = backbone
+        self.pretrained = pretrained
+
         super().__init__(
             label_info=label_info,
             optimizer=optimizer,
@@ -188,74 +178,22 @@ class MobileNetV3ForMultilabelCls(OTXMultilabelClsModel):
         return model
 
     def _build_model(self, num_classes: int) -> nn.Module:
+        backbone = TimmBackbone(backbone=self.backbone, pretrained=self.pretrained)
         return ImageClassifier(
-            backbone=OTXMobileNetV3(mode=self.mode),
+            backbone=backbone,
             neck=GlobalAveragePooling(dim=2),
-            head=MultiLabelNonLinearClsHead(
+            head=MultiLabelLinearClsHead(
                 num_classes=num_classes,
-                in_channels=960,
-                hid_channels=1280,
+                in_channels=backbone.num_features,
                 normalized=True,
                 scale=7.0,
-                activation_callable=nn.PReLU(),
                 loss=AsymmetricAngularLossWithIgnore(gamma_pos=0.0, gamma_neg=1.0, reduction="sum"),
             ),
         )
 
     def load_from_otx_v1_ckpt(self, state_dict: dict, add_prefix: str = "model.") -> dict:
         """Load the previous OTX ckpt according to OTX2.0."""
-        return OTXv1Helper.load_cls_mobilenet_v3_ckpt(state_dict, "multilabel", add_prefix)
-
-    def _customize_inputs(self, inputs: MultilabelClsBatchDataEntity) -> dict[str, Any]:
-        if self.training:
-            mode = "loss"
-        elif self.explain_mode:
-            mode = "explain"
-        else:
-            mode = "predict"
-
-        return {
-            "images": inputs.stacked_images,
-            "labels": torch.stack(inputs.labels),
-            "imgs_info": inputs.imgs_info,
-            "mode": mode,
-        }
-
-    def _customize_outputs(
-        self,
-        outputs: Any,  # noqa: ANN401
-        inputs: MultilabelClsBatchDataEntity,
-    ) -> MultilabelClsBatchPredEntity | OTXBatchLossEntity:
-        if self.training:
-            return OTXBatchLossEntity(loss=outputs)
-
-        # To list, batch-wise
-        logits = outputs if isinstance(outputs, torch.Tensor) else outputs["logits"]
-        scores = torch.unbind(logits, 0)
-
-        return MultilabelClsBatchPredEntity(
-            batch_size=inputs.batch_size,
-            images=inputs.images,
-            imgs_info=inputs.imgs_info,
-            scores=scores,
-            labels=logits.argmax(-1, keepdim=True).unbind(0),
-        )
-
-    @property
-    def _exporter(self) -> OTXModelExporter:
-        """Creates OTXModelExporter object that can export the model."""
-        return OTXNativeModelExporter(
-            task_level_export_parameters=self._export_parameters,
-            input_size=(1, 3, 224, 224),
-            mean=(123.675, 116.28, 103.53),
-            std=(58.395, 57.12, 57.375),
-            resize_mode="standard",
-            pad_value=0,
-            swap_rgb=False,
-            via_onnx=False,
-            onnx_export_configuration=None,
-            output_names=["logits", "feature_vector", "saliency_map"] if self.explain_mode else None,
-        )
+        return OTXv1Helper.load_cls_effnet_v2_ckpt(state_dict, "multilabel", add_prefix)
 
     def forward_explain(self, inputs: MultilabelClsBatchDataEntity) -> MultilabelClsBatchPredEntity:
         """Model forward explain function."""
@@ -271,7 +209,7 @@ class MobileNetV3ForMultilabelCls(OTXMultilabelClsModel):
             feature_vector=outputs["feature_vector"],
         )
 
-    def forward_for_tracing(self, image: Tensor) -> Tensor | dict[str, Tensor]:
+    def forward_for_tracing(self, image: torch.Tensor) -> torch.Tensor | dict[str, torch.Tensor]:
         """Model forward function used for the model tracing during model exportation."""
         if self.explain_mode:
             return self.model(images=image, mode="explain")
@@ -279,21 +217,24 @@ class MobileNetV3ForMultilabelCls(OTXMultilabelClsModel):
         return self.model(images=image, mode="tensor")
 
 
-class MobileNetV3ForHLabelCls(OTXHlabelClsModel):
-    """MobileNetV3 Model for hierarchical label classification task."""
+class TimmModelForHLabelCls(OTXHlabelClsModel):
+    """EfficientNetV2 Model for hierarchical label classification task."""
 
     label_info: HLabelInfo
 
     def __init__(
         self,
         label_info: HLabelInfo,
-        mode: Literal["large", "small"] = "large",
+        backbone: TimmModelType = "efficientnetv2_s_21k",
+        pretrained: bool = True,
         optimizer: OptimizerCallable = DefaultOptimizerCallable,
         scheduler: LRSchedulerCallable | LRSchedulerListCallable = DefaultSchedulerCallable,
         metric: MetricCallable = HLabelClsMetricCallble,
         torch_compile: bool = False,
     ) -> None:
-        self.mode = mode
+        self.backbone = backbone
+        self.pretrained = pretrained
+
         super().__init__(
             label_info=label_info,
             optimizer=optimizer,
@@ -320,14 +261,12 @@ class MobileNetV3ForHLabelCls(OTXHlabelClsModel):
         return model
 
     def _build_model(self, head_config: dict) -> nn.Module:
-        if not isinstance(self.label_info, HLabelInfo):
-            raise TypeError(self.label_info)
-
+        backbone = TimmBackbone(backbone=self.backbone, pretrained=self.pretrained)
         return ImageClassifier(
-            backbone=OTXMobileNetV3(mode=self.mode),
+            backbone=backbone,
             neck=GlobalAveragePooling(dim=2),
-            head=HierarchicalNonLinearClsHead(
-                in_channels=960,
+            head=HierarchicalLinearClsHead(
+                in_channels=backbone.num_features,
                 multiclass_loss=nn.CrossEntropyLoss(),
                 multilabel_loss=AsymmetricAngularLossWithIgnore(gamma_pos=0.0, gamma_neg=1.0, reduction="sum"),
                 **head_config,
@@ -336,82 +275,7 @@ class MobileNetV3ForHLabelCls(OTXHlabelClsModel):
 
     def load_from_otx_v1_ckpt(self, state_dict: dict, add_prefix: str = "model.") -> dict:
         """Load the previous OTX ckpt according to OTX2.0."""
-        return OTXv1Helper.load_cls_mobilenet_v3_ckpt(state_dict, "hlabel", add_prefix)
-
-    def _customize_inputs(self, inputs: HlabelClsBatchDataEntity) -> dict[str, Any]:
-        if self.training:
-            mode = "loss"
-        elif self.explain_mode:
-            mode = "explain"
-        else:
-            mode = "predict"
-
-        return {
-            "images": inputs.stacked_images,
-            "labels": torch.stack(inputs.labels),
-            "imgs_info": inputs.imgs_info,
-            "mode": mode,
-        }
-
-    def _customize_outputs(
-        self,
-        outputs: Any,  # noqa: ANN401
-        inputs: HlabelClsBatchDataEntity,
-    ) -> HlabelClsBatchPredEntity | OTXBatchLossEntity:
-        if self.training:
-            return OTXBatchLossEntity(loss=outputs)
-
-        # To list, batch-wise
-        if isinstance(outputs, dict):
-            scores = outputs["scores"]
-            labels = outputs["labels"]
-        else:
-            scores = outputs
-            labels = outputs.argmax(-1, keepdim=True)
-
-        return HlabelClsBatchPredEntity(
-            batch_size=inputs.batch_size,
-            images=inputs.images,
-            imgs_info=inputs.imgs_info,
-            scores=scores,
-            labels=labels,
-        )
-
-    def _convert_pred_entity_to_compute_metric(
-        self,
-        preds: HlabelClsBatchPredEntity,
-        inputs: HlabelClsBatchDataEntity,
-    ) -> MetricInput:
-        hlabel_info: HLabelInfo = self.label_info  # type: ignore[assignment]
-
-        _labels = torch.stack(preds.labels) if isinstance(preds.labels, list) else preds.labels
-        _scores = torch.stack(preds.scores) if isinstance(preds.scores, list) else preds.scores
-        if hlabel_info.num_multilabel_classes > 0:
-            preds_multiclass = _labels[:, : hlabel_info.num_multiclass_heads]
-            preds_multilabel = _scores[:, hlabel_info.num_multiclass_heads :]
-            pred_result = torch.cat([preds_multiclass, preds_multilabel], dim=1)
-        else:
-            pred_result = _labels
-        return {
-            "preds": pred_result,
-            "target": torch.stack(inputs.labels),
-        }
-
-    @property
-    def _exporter(self) -> OTXModelExporter:
-        """Creates OTXModelExporter object that can export the model."""
-        return OTXNativeModelExporter(
-            task_level_export_parameters=self._export_parameters,
-            input_size=(1, 3, 224, 224),
-            mean=(123.675, 116.28, 103.53),
-            std=(58.395, 57.12, 57.375),
-            resize_mode="standard",
-            pad_value=0,
-            swap_rgb=False,
-            via_onnx=False,
-            onnx_export_configuration=None,
-            output_names=["logits", "feature_vector", "saliency_map"] if self.explain_mode else None,
-        )
+        return OTXv1Helper.load_cls_effnet_v2_ckpt(state_dict, "hlabel", add_prefix)
 
     def forward_explain(self, inputs: HlabelClsBatchDataEntity) -> HlabelClsBatchPredEntity:
         """Model forward explain function."""
@@ -427,7 +291,7 @@ class MobileNetV3ForHLabelCls(OTXHlabelClsModel):
             feature_vector=outputs["feature_vector"],
         )
 
-    def forward_for_tracing(self, image: Tensor) -> Tensor | dict[str, Tensor]:
+    def forward_for_tracing(self, image: torch.Tensor) -> torch.Tensor | dict[str, torch.Tensor]:
         """Model forward function used for the model tracing during model exportation."""
         if self.explain_mode:
             return self.model(images=image, mode="explain")
