+++ src/otx/core/model/classification.py
@@ -4,12 +4,13 @@
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING, Any, Literal
 
 import numpy as np
 import torch
 from torch import Tensor
 
+from otx.core.data.entity.base import OTXBatchLossEntity
 from otx.core.data.entity.classification import (
     HlabelClsBatchDataEntity,
     HlabelClsBatchPredEntity,
@@ -18,6 +19,8 @@ from otx.core.data.entity.classification import (
     MultilabelClsBatchDataEntity,
     MultilabelClsBatchPredEntity,
 )
+from otx.core.exporter.base import OTXModelExporter
+from otx.core.exporter.native import OTXNativeModelExporter
 from otx.core.metrics import MetricInput
 from otx.core.metrics.accuracy import (
     HLabelClsMetricCallble,
@@ -28,6 +31,7 @@ from otx.core.model.base import DefaultOptimizerCallable, DefaultSchedulerCallab
 from otx.core.schedulers import LRSchedulerListCallable
 from otx.core.types.export import TaskLevelExportParameters
 from otx.core.types.label import HLabelInfo, LabelInfo, LabelInfoTypes
+from otx.core.types.task import OTXTrainType
 
 if TYPE_CHECKING:
     from lightning.pytorch.cli import LRSchedulerCallable, OptimizerCallable
@@ -46,7 +50,10 @@ class OTXMulticlassClsModel(OTXModel[MulticlassClsBatchDataEntity, MulticlassCls
         scheduler: LRSchedulerCallable | LRSchedulerListCallable = DefaultSchedulerCallable,
         metric: MetricCallable = MultiClassClsMetricCallable,
         torch_compile: bool = False,
+        train_type: Literal[OTXTrainType.SUPERVISED, OTXTrainType.SEMI_SUPERVISED] = OTXTrainType.SUPERVISED,
     ) -> None:
+        self.train_type = train_type
+
         super().__init__(
             label_info=label_info,
             optimizer=optimizer,
@@ -56,6 +63,88 @@ class OTXMulticlassClsModel(OTXModel[MulticlassClsBatchDataEntity, MulticlassCls
         )
         self.image_size = (1, 3, 224, 224)
 
+    def _customize_inputs(self, inputs: MulticlassClsBatchDataEntity) -> dict[str, Any]:
+        if self.training:
+            mode = "loss"
+        elif self.explain_mode:
+            mode = "explain"
+        else:
+            mode = "predict"
+
+        if self.train_type == OTXTrainType.SEMI_SUPERVISED and isinstance(inputs, dict):
+            # When used with an unlabeled dataset, it comes in as a dict.
+            images = {key: inputs[key].images for key in inputs}
+            labels = {key: torch.cat(inputs[key].labels, dim=0) for key in inputs}
+            imgs_info = {key: inputs[key].imgs_info for key in inputs}
+            return {
+                "images": images,
+                "labels": labels,
+                "imgs_info": imgs_info,
+                "mode": mode,
+            }
+
+        return {
+            "images": inputs.stacked_images,
+            "labels": torch.cat(inputs.labels, dim=0),
+            "imgs_info": inputs.imgs_info,
+            "mode": mode,
+        }
+
+    def _customize_outputs(
+        self,
+        outputs: Any,  # noqa: ANN401
+        inputs: MulticlassClsBatchDataEntity,
+    ) -> MulticlassClsBatchPredEntity | OTXBatchLossEntity:
+        if self.training:
+            return OTXBatchLossEntity(loss=outputs)
+
+        if self.explain_mode:
+            return MulticlassClsBatchPredEntity(
+                batch_size=inputs.batch_size,
+                images=inputs.images,
+                imgs_info=inputs.imgs_info,
+                scores=outputs["scores"],
+                labels=outputs["labels"],
+                saliency_map=outputs["saliency_map"],
+                feature_vector=outputs["feature_vector"],
+            )
+
+        # To list, batch-wise
+        logits = outputs if isinstance(outputs, torch.Tensor) else outputs["logits"]
+        scores = torch.unbind(logits, 0)
+        preds = logits.argmax(-1, keepdim=True).unbind(0)
+
+        return MulticlassClsBatchPredEntity(
+            batch_size=inputs.batch_size,
+            images=inputs.stacked_images,
+            imgs_info=inputs.imgs_info,
+            scores=scores,
+            labels=preds,
+        )
+
+    def training_step(self, batch: MulticlassClsBatchDataEntity, batch_idx: int) -> Tensor:
+        """Performs a single training step on a batch of data."""
+        loss = super().training_step(batch, batch_idx)
+        # Collect metrics related to Semi-SL Training.
+        if self.train_type == OTXTrainType.SEMI_SUPERVISED:
+            if hasattr(self.model.head, "unlabeled_coef"):
+                self.log(
+                    "train/unlabeled_coef",
+                    self.model.head.unlabeled_coef,
+                    on_step=True,
+                    on_epoch=False,
+                    prog_bar=True,
+                )
+            if hasattr(self.model.head, "num_pseudo_label"):
+                self.log(
+                    "train/num_pseudo_label",
+                    self.model.head.num_pseudo_label,
+                    on_step=True,
+                    on_epoch=False,
+                    prog_bar=True,
+                )
+        return loss
+
     @property
     def _export_parameters(self) -> TaskLevelExportParameters:
         """Defines parameters required to export a particular model implementation."""
@@ -64,6 +153,22 @@ class OTXMulticlassClsModel(OTXModel[MulticlassClsBatchDataEntity, MulticlassCls
             task_type="classification",
             multilabel=False,
             hierarchical=False,
+        )
+
+    @property
+    def _exporter(self) -> OTXModelExporter:
+        """Creates OTXModelExporter object that can export the model."""
+        return OTXNativeModelExporter(
+            task_level_export_parameters=self._export_parameters,
+            input_size=self.image_size,
+            mean=(123.675, 116.28, 103.53),
+            std=(58.395, 57.12, 57.375),
+            resize_mode="standard",
+            pad_value=0,
+            swap_rgb=False,
+            via_onnx=False,
+            onnx_export_configuration=None,
+            output_names=["logits", "feature_vector", "saliency_map"] if self.explain_mode else None,
         )
 
     def _convert_pred_entity_to_compute_metric(
@@ -78,11 +183,18 @@ class OTXMulticlassClsModel(OTXModel[MulticlassClsBatchDataEntity, MulticlassCls
             "target": target,
         }
 
+    def _reset_prediction_layer(self, num_classes: int) -> None:
+        return
+
     def get_dummy_input(self, batch_size: int = 1) -> MulticlassClsBatchDataEntity:
         """Returns a dummy input for classification model."""
         images = [torch.rand(*self.image_size[1:]) for _ in range(batch_size)]
         labels = [torch.LongTensor([0])] * batch_size
         return MulticlassClsBatchDataEntity(batch_size, images, [], labels=labels)
+
+    def forward_for_tracing(self, image: Tensor) -> Tensor | dict[str, Tensor]:
+        """Model forward function used for the model tracing during model exportation."""
+        return self.model(images=image)
 
 
 ### NOTE, currently, although we've made the separate Multi-cls, Multi-label classes
@@ -109,6 +221,52 @@ class OTXMultilabelClsModel(OTXModel[MultilabelClsBatchDataEntity, MultilabelCls
         )
         self.image_size = (1, 3, 224, 224)
 
+    def _customize_inputs(self, inputs: MultilabelClsBatchDataEntity) -> dict[str, Any]:
+        if self.training:
+            mode = "loss"
+        elif self.explain_mode:
+            mode = "explain"
+        else:
+            mode = "predict"
+
+        return {
+            "images": inputs.stacked_images,
+            "labels": torch.stack(inputs.labels),
+            "imgs_info": inputs.imgs_info,
+            "mode": mode,
+        }
+
+    def _customize_outputs(
+        self,
+        outputs: Any,  # noqa: ANN401
+        inputs: MultilabelClsBatchDataEntity,
+    ) -> MultilabelClsBatchPredEntity | OTXBatchLossEntity:
+        if self.training:
+            return OTXBatchLossEntity(loss=outputs)
+
+        if self.explain_mode:
+            return MultilabelClsBatchPredEntity(
+                batch_size=inputs.batch_size,
+                images=inputs.images,
+                imgs_info=inputs.imgs_info,
+                scores=outputs["scores"],
+                labels=outputs["labels"],
+                saliency_map=outputs["saliency_map"],
+                feature_vector=outputs["feature_vector"],
+            )
+
+        # To list, batch-wise
+        logits = outputs if isinstance(outputs, torch.Tensor) else outputs["logits"]
+        scores = torch.unbind(logits, 0)
+
+        return MultilabelClsBatchPredEntity(
+            batch_size=inputs.batch_size,
+            images=inputs.images,
+            imgs_info=inputs.imgs_info,
+            scores=scores,
+            labels=logits.argmax(-1, keepdim=True).unbind(0),
+        )
+
     @property
     def _export_parameters(self) -> TaskLevelExportParameters:
         """Defines parameters required to export a particular model implementation."""
@@ -118,6 +276,22 @@ class OTXMultilabelClsModel(OTXModel[MultilabelClsBatchDataEntity, MultilabelCls
             multilabel=True,
             hierarchical=False,
             confidence_threshold=0.5,
+        )
+
+    @property
+    def _exporter(self) -> OTXModelExporter:
+        """Creates OTXModelExporter object that can export the model."""
+        return OTXNativeModelExporter(
+            task_level_export_parameters=self._export_parameters,
+            input_size=self.image_size,
+            mean=(123.675, 116.28, 103.53),
+            std=(58.395, 57.12, 57.375),
+            resize_mode="standard",
+            pad_value=0,
+            swap_rgb=False,
+            via_onnx=False,
+            onnx_export_configuration=None,
+            output_names=["logits", "feature_vector", "saliency_map"] if self.explain_mode else None,
         )
 
     def _convert_pred_entity_to_compute_metric(
@@ -132,7 +306,7 @@ class OTXMultilabelClsModel(OTXModel[MultilabelClsBatchDataEntity, MultilabelCls
 
     def forward_for_tracing(self, image: Tensor) -> Tensor | dict[str, Tensor]:
         """Model forward function used for the model tracing during model exportation."""
-        return self.model.forward(image, mode="tensor")
+        return self.model.forward(image)
 
     def get_dummy_input(self, batch_size: int = 1) -> MultilabelClsBatchDataEntity:
         """Returns a dummy input for classification OV model."""
@@ -143,6 +317,8 @@ class OTXMultilabelClsModel(OTXModel[MultilabelClsBatchDataEntity, MultilabelCls
 
 class OTXHlabelClsModel(OTXModel[HlabelClsBatchDataEntity, HlabelClsBatchPredEntity]):
     """H-label classification models used in OTX."""
+
+    label_info: HLabelInfo
 
     def __init__(
         self,
@@ -161,6 +337,56 @@ class OTXHlabelClsModel(OTXModel[HlabelClsBatchDataEntity, HlabelClsBatchPredEnt
         )
         self.image_size = (1, 3, 224, 224)
 
+    def _customize_inputs(self, inputs: HlabelClsBatchDataEntity) -> dict[str, Any]:
+        if self.training:
+            mode = "loss"
+        elif self.explain_mode:
+            mode = "explain"
+        else:
+            mode = "predict"
+
+        return {
+            "images": inputs.stacked_images,
+            "labels": torch.stack(inputs.labels),
+            "imgs_info": inputs.imgs_info,
+            "mode": mode,
+        }
+
+    def _customize_outputs(
+        self,
+        outputs: Any,  # noqa: ANN401
+        inputs: HlabelClsBatchDataEntity,
+    ) -> HlabelClsBatchPredEntity | OTXBatchLossEntity:
+        if self.training:
+            return OTXBatchLossEntity(loss=outputs)
+
+        # To list, batch-wise
+        if isinstance(outputs, dict):
+            scores = outputs["scores"]
+            labels = outputs["labels"]
+        else:
+            scores = outputs
+            labels = outputs.argmax(-1, keepdim=True)
+
+        if self.explain_mode:
+            return HlabelClsBatchPredEntity(
+                batch_size=inputs.batch_size,
+                images=inputs.images,
+                imgs_info=inputs.imgs_info,
+                scores=scores,
+                labels=labels,
+                saliency_map=outputs["saliency_map"],
+                feature_vector=outputs["feature_vector"],
+            )
+
+        return HlabelClsBatchPredEntity(
+            batch_size=inputs.batch_size,
+            images=inputs.images,
+            imgs_info=inputs.imgs_info,
+            scores=scores,
+            labels=labels,
+        )
+
     @property
     def _export_parameters(self) -> TaskLevelExportParameters:
         """Defines parameters required to export a particular model implementation."""
@@ -172,6 +398,22 @@ class OTXHlabelClsModel(OTXModel[HlabelClsBatchDataEntity, HlabelClsBatchPredEnt
             confidence_threshold=0.5,
         )
 
+    @property
+    def _exporter(self) -> OTXModelExporter:
+        """Creates OTXModelExporter object that can export the model."""
+        return OTXNativeModelExporter(
+            task_level_export_parameters=self._export_parameters,
+            input_size=self.image_size,
+            mean=(123.675, 116.28, 103.53),
+            std=(58.395, 57.12, 57.375),
+            resize_mode="standard",
+            pad_value=0,
+            swap_rgb=False,
+            via_onnx=False,
+            onnx_export_configuration=None,
+            output_names=["logits", "feature_vector", "saliency_map"] if self.explain_mode else None,
+        )
+
     def _convert_pred_entity_to_compute_metric(
         self,
         preds: HlabelClsBatchPredEntity,
@@ -179,12 +421,14 @@ class OTXHlabelClsModel(OTXModel[HlabelClsBatchDataEntity, HlabelClsBatchPredEnt
     ) -> MetricInput:
         hlabel_info: HLabelInfo = self.label_info  # type: ignore[assignment]
 
+        _labels = torch.stack(preds.labels) if isinstance(preds.labels, list) else preds.labels
+        _scores = torch.stack(preds.scores) if isinstance(preds.scores, list) else preds.scores
         if hlabel_info.num_multilabel_classes > 0:
-            preds_multiclass = torch.stack(preds.labels)[:, : hlabel_info.num_multiclass_heads]
-            preds_multilabel = torch.stack(preds.scores)[:, hlabel_info.num_multiclass_heads :]
+            preds_multiclass = _labels[:, : hlabel_info.num_multiclass_heads]
+            preds_multilabel = _scores[:, hlabel_info.num_multiclass_heads :]
             pred_result = torch.cat([preds_multiclass, preds_multilabel], dim=1)
         else:
-            pred_result = torch.stack(preds.labels)
+            pred_result = _labels
         return {
             "preds": pred_result,
             "target": torch.stack(inputs.labels),
@@ -202,6 +446,10 @@ class OTXHlabelClsModel(OTXModel[HlabelClsBatchDataEntity, HlabelClsBatchPredEnt
         images = [torch.rand(*self.image_size[1:]) for _ in range(batch_size)]
         labels = [torch.LongTensor([0])] * batch_size
         return HlabelClsBatchDataEntity(batch_size, images, [], labels=labels)
+
+    def forward_for_tracing(self, image: Tensor) -> Tensor | dict[str, Tensor]:
+        """Model forward function used for the model tracing during model exportation."""
+        return self.model(images=image)
 
 
 class OVMulticlassClassificationModel(
