+++ src/otx/recipe/classification/multi_class_cls/efficientnet_v2_semisl.yaml
@@ -1,7 +1,9 @@
 model:
-  class_path: otx.algo.classification.efficientnet_v2.EfficientNetV2ForMulticlassClsSemiSL
+  class_path: otx.algo.classification.timm_model.TimmModelForMulticlassCls
   init_args:
     label_info: 1000
+    backbone: efficientnetv2_s_21k
+    train_type: SEMI_SUPERVISED
 
     optimizer:
       class_path: torch.optim.SGD
