+++ tests/unit/algo/callbacks/test_unlabeled_loss_warmup.py
@@ -3,12 +3,12 @@ import math
 import pytest
 from lightning import Trainer
 from otx.algo.callbacks.unlabeled_loss_warmup import UnlabeledLossWarmUpCallback
-from otx.algo.classification.efficientnet import EfficientNetForMulticlassClsSemiSL
+from otx.algo.classification.efficientnet import EfficientNetForMulticlassCls
 
 
 @pytest.fixture()
 def fxt_semisl_model():
-    return EfficientNetForMulticlassClsSemiSL(10)
+    return EfficientNetForMulticlassCls(10, train_type="SEMI_SUPERVISED")
 
 
 def test_unlabeled_loss_warmup_callback(mocker, fxt_semisl_model):
